feat: K8s ConfigMap/STS/Deployment YAML 查看编辑 + Pod 终端 + 容器模块增强
1. ConfigMap: 新增 YAML 查看/编辑（GET+PUT /yaml），前端详情页支持在线编辑
2. StatefulSet: 新增 YAML 查看/编辑，列表支持导航到详情
3. Deployment: 新增 PUT /yaml 更新端点
4. 前端: ResourceYamlEditor 组件（textarea 编辑 + 查看模式切换）
5. 前端: ConfigMapDetailView YAML 标签页 + 列表页面加行跳转
6. 容器路由: 新增 configmap-detail 子路由
7. StatefulSet/Secret/PVC 后端: 补齐 Controller/Service/DTO
8. Pod 终端: PodShellTerminal WebSocket 交互终端
9. 集群管理: Node 详情/列表增强，Cluster 管理完善
10. API 完善: event/node/pod/service 前端 API 补充
11. 容器模块: AutoConfiguration + WsConfig + ClusterService 增强

+++ backend/container-core/src/main/java/com/hfwas/devops/container/dto/NodeDetailVO.java
@@ -64,6 +64,6 @@ public class NodeDetailVO {
     @Data
     public static class NodeImageVO {
         private String name;
-        private int sizeBytes;
+        private long sizeBytes;
     }
 }
